// src/modules/notifications/enums/email-type.enum.ts
export enum EmailType {
  WELCOME = 'WELCOME',
  LIKE_SENT = 'LIKE_SENT',
  NEW_MATCH = 'NEW_MATCH',
  NEW_MESSAGE = 'NEW_MESSAGE',
  CREDIT_UPDATE = 'CREDIT_UPDATE',
  CREDITS_PURCHASED = 'CREDITS_PURCHASED',
  SYSTEM_ALERT = 'SYSTEM_ALERT',
  BUNDLE_EXPIRY_WARNING = 'BUNDLE_EXPIRY_WARNING',
  LIKES_EXPIRED = 'LIKES_EXPIRED',
  IDENTITY_ADDED = 'IDENTITY_ADDED',
}